fix: 린트, 빌드 에러 수정
