+++ hanarowa/src/app/admin/lesson/manage/[id]/page.tsx
@@ -242,6 +242,8 @@ function DetailForm() {
             />
             {formData.lessonImage ? (
               <div className='relative'>
+                {/* 나중에 Next Image로 바꾸세요. */}
+                {/* eslint-disable-next-line @next/next/no-img-element */}
                 <img
                   src={URL.createObjectURL(formData.lessonImage)}
                   alt='업로드된 이미지'
@@ -259,6 +261,8 @@ function DetailForm() {
               </div>
             ) : initial?.imageUrl ? (
               <div className='relative'>
+                {/* 나중에 Next Image로 바꾸세요. */}
+                {/* eslint-disable-next-line @next/next/no-img-element */}
                 <img
                   src={initial.imageUrl}
                   alt='기존 이미지'
