+++ hanarowa/src/app/auth/login/actions.ts
@@ -36,6 +36,7 @@ export async function login(
     // login api
     return { success: true, message: '로그인 성공!' };
   } catch (err) {
+    console.log(err);
     return {
       success: false,
       message: '이메일 또는 비밀번호를 확인해주세요.',
