+++ hanarowa/src/components/lesson/StatusTab.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { motion } from 'framer-motion';
 import React from 'react';
